Modified LSL script for the RegEnrol booth. Removed stray character which made "install.xml" invalid XML. Added ability to customize frame dimensions for Presenter in Moodle. Added ability to name each slide in the Presenter, and updated linker/LSL scripts. Added Presenter translation script to repository. Filtered list of avatars on Distributor page -- only shows avatars whose Moodle account is enrolled in the given course. Fixed "SLOODLE Modules" page, which showed error messages for unrecognised module types. Resolved Issue 74 by altering the "print_header" function of the base module view class. Update LSL scripts to fix bugs in WebIntercom, Vending Machine, and Set recycle bin.
Added "sloodle_clean_for_output" and "sloodle_clean_for_db" functions to "lib/general.php", for cleaning text. It is HIGHLY ADVISABLE to use the output function when sending text for SL, as this will ensure the HTML entities etc. are correctly rendered for Unicode, not ASCII.

Made WebIntercom, glossary and quiz compatible with Unicode characters.
Fixed issue preventing Distributor from handling objects with HTML characters.
Fixed broken avatar link when visiting a LoginZone with an account that already has an avatar linked to it.
Updated SLOODLE Choice to show the same results in-world as will be shown in Moodle.