FIX:  In the older version, the items in the vending machine  were listed out of order in the menu.  This has now been fixed, and all items are listed in the correct order
ENHANCEMENT:  Before, several hovertext messages were displayed in dark blue. This was hard to read, so I replaced with an easy to read bright pink - which seams to be easily readable in world over most backgrounds.

ENHANCEMENT:  I have also added an additional meni choice: "web".  This allows the user, or teacher, to jump to the MOODLE distributer page for the vending machine.  This can be useful in situations where the teacher or students needs quick access to the web distributer, and doesnt want to hunt for which activity is connected to the distributer they are currently using.  Good for situations where numerous vending machines exist in the same course.

FIX: In the previous vending machine, we used "<<" and ">>" for the previous and next pages.  It turns out that the "<<" caused a strange bug in some situations, possibily being mistaken by the lsl code a bitshifter operator in lsl - thus causing menu items to throw strange errors.  I've therefore changed the text labels for the next and previous pages to "NEXT" and "PREVIOUS"

FIX:  I also found that if you configure the vending machine using a sloodle_config, then delete the sloodle_config from the contents, the menu still displays but reports incorrect contents with negative button labels.  The problem was that the page number went below zero.  I've therefore added code that checks against a negative value, and resets the page number to zero.  This seams to fix the error.

We hope you enjoy the new updated vending machine with the new sorted menu, and the web button!