This branch has the beta HQ module which has an LSL API in it.
Once stable, will be merged into the trunk.

New releases such as Sloodle 1.1 should be a tag.
This is not 1.1, but instead a branch that will be merged back into the trunk later. When ready to release 1.1, we tag the trunk.  

ALL work on the api should go in this branch, until we are ready to merge back into the trunk