Removed fetching-ahead of questions to save memory, so that the quiz only loads one question at a time from the server. This introduces a slight delay between answering a question and asking the next one, as the script has to talk to the server and get the next question. But this turns out to work pretty well UI-wise anyway, because it focusses attention on the feedback for the question just answers before moving on to the next thing.
Also removed the dummy repeat_quiz state, in the hope of saving some memory.

If we still get stack-collisions after this, the next move will be to separate out the feedback into a separate call, so instead of loading in all the feedback for all the possible responses, we just load the feedback that we need for option that the user actually chose.