Implemented ability to use portable course and module identifiers in object descriptions (in-world), allowing configured objects to be transferred to different grids (e.g. via OAR file) alongside a course backup in Moodle.
One particular weakness is that it requires the destination course (in Moodle) to have the same portable identifier. It would be useful to be able to automatically modify the OAR file prior to loading so that a different course identifier can be included. Alternatively, it should be possible to re-configure the course identifier of all objects immediately after they've been rezzed.