Removed some old SL Map code which shouldn't have been re-added. Removed out-dated Awards database code which was causing errors. Removed unnecessarily duplicated code in Awards view script which was masking some core default functionality. Added missing call to core setup functionality in Awards view script (necessary objects were not being populated, sometimes resulting in errors/warnings). Removed out-dated default values from Awards module update form. Resolve several errors in the game view tab on the Awards view page. Removed some redundant code from awards object.
Did as much as I could to update and fix the Awards backup/restore code. Seems to work for the Awards system itself, players, and games. I don't think it works for transactions due to the inconsistent (and incorrect) usage of the sloodleid field. That may need quite a bit of work to fix.